fix(platform:menu): docs fixes for Stackblitz
also made some properties and methods hidden from documentation pages
