+++ apps/docs/src/app/platform/component-docs/platform-menu/platform-menu-docs.component.html
@@ -3,7 +3,7 @@
 </fd-docs-section-title>
 <description>This example shows Menu with a <code>fd-dropdown-control</code> as invoker</description>
 <component-example [name]="'ex1'">
-    <fdp-menu-example></fdp-menu-example>
+    <fdp-platform-menu-basic-example></fdp-platform-menu-basic-example>
 </component-example>
 <code-example [exampleFiles]="menuBasic"></code-example>
 
@@ -14,7 +14,7 @@
 </fd-docs-section-title>
 <description>Each menu item can have an optional separator shown as a line beneath the item.</description>
 <component-example [name]="'ex2'">
-    <fdp-menu-separator-example></fdp-menu-separator-example>
+    <fdp-platform-menu-separator-example></fdp-platform-menu-separator-example>
 </component-example>
 <code-example [exampleFiles]="menuSeparator"></code-example>
 
@@ -28,7 +28,7 @@
     achieved by using the <code>icon</code> and <code>secondaryIcon</code> fields while entering the data.
 </description>
 <component-example [name]="'ex3'">
-    <fdp-menu-icons-example></fdp-menu-icons-example>
+    <fdp-platform-menu-icons-example></fdp-platform-menu-icons-example>
 </component-example>
 <code-example [exampleFiles]="menuAddon"></code-example>
 
@@ -41,7 +41,7 @@
     This provides an additional grouping of some items for better readability.
 </description>
 <component-example [name]="'ex4'">
-    <fdp-menu-group-example></fdp-menu-group-example>
+    <fdp-platform-menu-group-example></fdp-platform-menu-group-example>
 </component-example>
 <code-example [exampleFiles]="menuGroup"></code-example>
 
@@ -52,7 +52,7 @@
 </fd-docs-section-title>
 <description>This example shows Menu with a <code>fd-popover-control</code> using button as the invoker</description>
 <component-example [name]="'ex5'">
-    <fdp-menu-complex-example></fdp-menu-complex-example>
+    <fdp-platform-menu-complex-example></fdp-platform-menu-complex-example>
 </component-example>
 <code-example [exampleFiles]="menuComplex"></code-example>
