+++ apps/docs/src/app/platform/component-docs/platform-menu/platform-menu-docs.component.ts
@@ -2,11 +2,25 @@ import { OnInit, Component } from '@angular/core';
 import { ExampleFile } from '../../../documentation/core-helpers/code-example/example-file';
 
 import * as platformBasicMenuSrc from '!raw-loader!./platform-menu-examples/platform-menu-basic-example.component.html';
+import * as platformBasicMenuTsCode from '!raw-loader!./platform-menu-examples/platform-menu-basic-example.component.ts';
+import * as platformBasicMenuScss from '!raw-loader!./platform-menu-examples/platform-menu-basic-example.component.scss';
+
 import * as platformMenuSeparatorSrc from '!raw-loader!./platform-menu-examples/platform-menu-separator-example.component.html';
+import * as platformMenuSeparatorTsCode from '!raw-loader!./platform-menu-examples/platform-menu-separator-example.component.ts';
+import * as platformMenuSeparatorScss from '!raw-loader!./platform-menu-examples/platform-menu-separator-example.component.scss';
+
 import * as platformMenuAddonSrc from '!raw-loader!./platform-menu-examples/platform-menu-icons-example.component.html';
+import * as platformMenuAddonTsCode from '!raw-loader!./platform-menu-examples/platform-menu-icons-example.component.ts';
+import * as platformMenuAddonScss from '!raw-loader!./platform-menu-examples/platform-menu-icons-example.component.scss';
+
 import * as platformMenuGroupSrc from '!raw-loader!./platform-menu-examples/platform-menu-group-example.component.html';
+import * as platformMenuGroupTsCode from '!raw-loader!./platform-menu-examples/platform-menu-group-example.component.ts';
+import * as platformMenuGroupScss from '!raw-loader!./platform-menu-examples/platform-menu-group-example.component.scss';
+
 import * as platformComplexMenuSrc from '!raw-loader!./platform-menu-examples/platform-menu-complex-example.component.html';
-import * as platformMenuTsCode from '!raw-loader!./platform-menu-examples/platform-menu-examples.component.ts';
+import * as platformComplexMenuTsCode from '!raw-loader!./platform-menu-examples/platform-menu-complex-example.component.ts';
+import * as platformComplexMenuScss from '!raw-loader!./platform-menu-examples/platform-menu-complex-example.component.scss';
+
 import { DocsSectionTitleComponent } from '../../../documentation/core-helpers/docs-section-title/docs-section-title.component';
 import { ActivatedRoute } from '@angular/router';
 
@@ -20,9 +34,13 @@ export class PlatformMenuDocsComponent implements OnInit {
             language: 'html',
             code: platformBasicMenuSrc,
             fileName: 'platform-menu-basic-example',
-            secondFile: 'platform-menu-examples',
-            typescriptFileCode: platformMenuTsCode,
-            component: 'PlatformBasicMenuExampleComponent'
+            scssFileCode: platformBasicMenuScss
+        },
+        {
+            language: 'typescript',
+            code: platformBasicMenuTsCode,
+            fileName: 'platform-menu-basic-example',
+            component: 'PlatformMenuBasicExampleComponent'
         }
     ];
 
@@ -31,8 +49,12 @@ export class PlatformMenuDocsComponent implements OnInit {
             language: 'html',
             code: platformMenuSeparatorSrc,
             fileName: 'platform-menu-separator-example',
-            secondFile: 'platform-menu-examples',
-            typescriptFileCode: platformMenuTsCode,
+            scssFileCode: platformMenuSeparatorScss
+        },
+        {
+            language: 'typescript',
+            code: platformMenuSeparatorTsCode,
+            fileName: 'platform-menu-separator-example',
             component: 'PlatformMenuSeparatorExampleComponent'
         }
     ];
@@ -42,8 +64,12 @@ export class PlatformMenuDocsComponent implements OnInit {
             language: 'html',
             code: platformMenuAddonSrc,
             fileName: 'platform-menu-icons-example',
-            secondFile: 'platform-menu-examples',
-            typescriptFileCode: platformMenuTsCode,
+            scssFileCode: platformMenuAddonScss
+        },
+        {
+            language: 'typescript',
+            code: platformMenuAddonTsCode,
+            fileName: 'platform-menu-icons-example',
             component: 'PlatformMenuIconsExampleComponent'
         }
     ];
@@ -53,8 +79,12 @@ export class PlatformMenuDocsComponent implements OnInit {
             language: 'html',
             code: platformMenuGroupSrc,
             fileName: 'platform-menu-group-example',
-            secondFile: 'platform-menu-examples',
-            typescriptFileCode: platformMenuTsCode,
+            scssFileCode: platformMenuGroupScss
+        },
+        {
+            language: 'typescript',
+            code: platformMenuGroupTsCode,
+            fileName: 'platform-menu-group-example',
             component: 'PlatformMenuGroupExampleComponent'
         }
     ];
@@ -64,9 +94,13 @@ export class PlatformMenuDocsComponent implements OnInit {
             language: 'html',
             code: platformComplexMenuSrc,
             fileName: 'platform-menu-complex-example',
-            secondFile: 'platform-menu-examples',
-            typescriptFileCode: platformMenuTsCode,
-            component: 'PlatformComplexMenuExampleComponent'
+            scssFileCode: platformComplexMenuScss
+        },
+        {
+            language: 'typescript',
+            code: platformComplexMenuTsCode,
+            fileName: 'platform-menu-complex-example',
+            component: 'PlatformMenuComplexExampleComponent'
         }
     ];
