+++ apps/docs/src/app/platform/component-docs/platform-menu/platform-menu-examples/platform-menu-group-example.component.scss
@@ -1,0 +1,1 @@
+@import '~fundamental-styles/dist/menu';
