+++ apps/docs/src/app/platform/component-docs/platform-menu/platform-menu-header/platform-menu-header.component.html
@@ -7,7 +7,7 @@
     Menu that can be used out-of-the-box with a tag <code>fdp-menu</code> with available attributes. Menu is navigable
     by pressing tab or arrow up/down, space and enter keys.
 </description>
-<import module="FdpMenuModule" path="@fundamental-ngx/platform"></import>
+<import module="MenuItem, MenuGroup" path="@fundamental-ngx/platform"></import>
 
 <fd-header-tabs></fd-header-tabs>
 <router-outlet></router-outlet>
